+++ Assets/Project/Scripts/GamePlayScene/Bottle/LifeNumberBottleController.cs
@@ -2,7 +2,7 @@
 using Project.Scripts.Utils.Definitions;
 using UnityEngine;
 
-namespace Project.Scripts.GamePlayScene.Panel
+namespace Project.Scripts.GamePlayScene.Bottle
 {
     /// <summary>
     /// ライフ付きのナンバーボトル
